Moved zipfile code to support.py; removed TutorialApp.py
Most of the useful code has now been extracted from TutorialApp.py

There's no longer any point keeping it around.  It can be accessed from
the commit history at a later date if necessary.
