--- code/tutorlib/TutorialApp.py
@@ -1,718 +1,0 @@
-
-## A Python Tutorial System
-## Copyright (C) 2009,2010  Peter Robinson <pjr@itee.uq.edu.au>
-##
-## This program is free software; you can redistribute it and/or
-## modify it under the terms of the GNU General Public License
-## as published by the Free Software Foundation; either version 2
-## of the License, or (at your option) any later version.
-##
-## This program is distributed in the hope that it will be useful,
-## but WITHOUT ANY WARRANTY; without even the implied warranty of
-## MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-## GNU General Public License for more details.
-##
-## You should have received a copy of the GNU General Public License
-## along with this program; if not, write to the Free Software
-## Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
-## 02110-1301, USA.
-
-## The MyPyTutor application
-
-
-import time
-import sys
-import os
-from io import StringIO
-import tkinter.messagebox
-import zipfile
-
-from tkinter import *
-import tkinter.filedialog
-
-from tutorlib.config.configuration import load_config, save_config, add_tutorial
-from tutorlib.gui.output import AnalysisOutput, TestOutput
-import tutorlib.Tutorial as tut_tutorial
-import tutorlib.TutorEditor as tut_editor
-from tutorlib.interface.tutorial_interface import TutorialInterface
-from tutorlib.gui.aboutdialog import TutAboutDialog
-from tutorlib.gui.helpdialog import HelpDialog
-from tutorlib.gui.feedbackdialog import FeedbackDialog
-from tutorlib.gui.font_chooser import FontChooser
-from tutorlib.gui.passworddialogs import ChangePasswordDialog, LoginDialog
-from tutorlib.gui.textdialog import TextDialog
-
-# Version number of MyPyTutor
-version_number = "3.00"
-
-# This application is intended to be run on the students machine.
-# Consequently, attempts have been made to hide how the program works.
-# First, the code is supplied in compiled form.
-# Secondly, comments have been written in such a way that help will not
-# show the comments.
-
-# unzip the zipfile (zf) to the given folder (path)
-
-def unzipit(zf, path):
-    z = zipfile.ZipFile(zf)
-    info = z.namelist()
-    if not os.path.exists(path):
-        os.mkdir(path)
-    for item in info:
-        if item.endswith('/') or item.endswith('\\'):
-            fulldir = os.path.join(path, item)
-            if not os.path.exists(fulldir):
-                os.mkdir(fulldir)
-        else:
-            flags = (z.getinfo(item).external_attr >> 16) & 0o777
-            text = z.read(item)
-            fullpath = os.path.join(path, item)
-            fd = open(fullpath, 'wb')
-            fd.write(text)
-            fd.close()
-            os.chmod(fullpath, flags)
-    z.close()
-
-
-# Toolbar is used to show the Hint button if there are any hints
-# for the problem and show to show the login status if MyPyTutor is
-# configured for online use.
-
-class Toolbar(Frame):
-    def __init__(self, parent, master=None):
-        Frame.__init__(self, master, bg='grey80')
-        self.parent = parent
-        self.hintbutton = Button(self, text="Next Hint",
-                                 command=self.parent.showhint)
-
-        self.bind("<Button-1>", self.button_pressed)
-        self.bind("<B1-Motion>", self.button_motion)
-        self.bind("<ButtonRelease-1>", self.button_released)
-        self.status_var = StringVar()
-
-    def button_pressed(self, event):
-        self.parent.allow_resize = False
-        self.y = event.y
-        self.start_len = self.parent.output_len
-        self.start_height = self.parent.output.winfo_height()
-
-    def button_motion(self, event):
-        # TODO: this is way too incestuous
-        # TODO: I'm also not convinced it's particularly useful
-        diff = (self.y - event.y)*self.start_len/self.start_height
-        if (self.parent.cfg.window_sizes.output + diff > 5 and
-                self.parent.cfg.window_sizes.problem - diff > 5):
-            self.parent.update_text_lengths(diff)
-
-    def button_released(self, event):
-        self.parent.allow_resize = True
-
-    def set_hints(self, flag):
-        if flag:
-            self.hintbutton.pack(side=LEFT, expand=1)
-        else:
-            self.hintbutton.pack_forget()
-
-    def enable_status(self):
-        self.status_var.set("Status: Not Logged In")
-        self.status = Label(self, textvariable=self.status_var, relief=SUNKEN)
-        self.status.pack(side=RIGHT, pady=3, ipady=2, padx=2, ipadx=2)
-
-    def set_login(self, user):
-        self.status_var.set("Status: Logged in as %s" % user)
-
-    def unset_login(self):
-        self.status_var.set("Status: Not Logged In")
-
-
-# The main application
-
-class TutorialApp():
-    def __init__(self, master=None):
-        #print "Starting..."
-        self.mpt_home = os.getcwd()
-        self.app_height = 0
-        self.allow_resize = False
-        self.URL = None
-
-
-        #print "done packing"
-        menubar = Menu(master)
-        sectionmenu = Menu(menubar, tearoff=0)
-        sectionmenu.add_command(label="Next Tutorial", accelerator="N",
-                                command=self.next_tutorial)
-        sectionmenu.add_command(label="Previous Tutorial",
-                                accelerator="P",
-                                command=self.previous_tutorial)
-        menubar.add_cascade(label="Problems", menu=sectionmenu)
-        self.sectionmenu = sectionmenu
-        self.sections = 0
-
-        # the configuration will not set the tutorial directories by default,
-        # so we need to do that
-        if not self.cfg.tutorials.default:
-            error = add_tutorial(self.cfg, window=self.master, as_default=True)
-
-            # just give up if this didn't work - the user can get it right
-            # next time they run this :)
-            if error:
-                return
-
-        self.current_tutorial = self.cfg.tutorials.default
-
-        options = getattr(self.cfg, self.current_tutorial)
-        self.tut_dir = options.tut_dir
-        self.ans_dir = options.ans_dir
-
-        self.setup_tutorial()
-        #print "done tutorial menu"
-        # If a URL is given then MyPyTutor has an online component and
-        # the following sets up a menu for online interaction
-        if self.URL:
-            wwwmenu = Menu(menubar, tearoff=0)
-            menubar.add_cascade(label="Online", menu=wwwmenu)
-            wwwmenu.add_command(label="Login",
-                                command=self.login)
-            wwwmenu.add_command(label="Logout",
-                                command=self.logout)
-            wwwmenu.add_command(label="Change Password",
-                                command=self.change_password)
-            wwwmenu.add_separator()
-            wwwmenu.add_command(label="Upload Problem Code",
-                                command=self.upload_answer)
-            wwwmenu.add_command(label="Download Problem Code",
-                                command=self.download_answer)
-            wwwmenu.add_separator()
-            wwwmenu.add_command(label="Submit Answer              F6",
-                                command=self.submit_answer)
-            wwwmenu.add_command(label="Show Submissions",
-                                command=self.show_submit)
-            master.bind("<F6>", self.submit_answer)
-            self.toolbar.enable_status()
-
-        self.tut.splash(self.URL, version_number)
-
-        optionsmenu = Menu(menubar, tearoff=0)
-        menubar.add_cascade(label="Preferences", menu=optionsmenu)
-        optionsmenu.add_command(label="Configure Tutor Fonts",
-                                command=self.configure_tut)
-        optionsmenu.add_command(label="Configure Tutorial Folder",
-                                command=self.configure_tut_dir)
-        optionsmenu.add_command(label="Configure Answers Folder",
-                                command=self.configure_answers_dir)
-        optionsmenu.add_separator()
-        optionsmenu.add_command(label="Set As Default Tutorial",
-                                command=self.set_default_tutorial)
-        optionsmenu.add_command(label="Add New Tutorial",
-                                command=self.add_new_tutorial)
-        optionsmenu.add_command(label="Remove Current Tutorial",
-                                command=self.remove_current_tutorial)
-        self.tut_choice = StringVar()
-        self.tut_choice.set(self.current_tutorial)
-        self.radio_menu = Menu(optionsmenu, tearoff=0)
-        optionsmenu.add_cascade(label="Change Tutorial", menu=self.radio_menu)
-
-        self.update_change_tutorial_menu()
-
-        feedbackmenu = Menu(menubar, tearoff=0)
-        menubar.add_cascade(label="MyPyTutor Feedback", menu=feedbackmenu)
-        feedbackmenu.add_command(label="Problem Feedback",
-                                 command=self.give_problem_feedback)
-        feedbackmenu.add_command(label="General Feedback",
-                                 command=self.give_general_feedback)
-
-        helpmenu = Menu(menubar, name='help', tearoff=0)
-        menubar.add_cascade(label="Help", menu=helpmenu)
-        helpmenu.add_command(label="About MyPyTutor",
-                             command=self.about)
-        helpmenu.add_command(label="Help",
-                             command=self.help)
-        master.config(menu=menubar)
-        #print "done all menus"
-        master.bind("n", self.next_tutorial)
-        master.bind("N", self.next_tutorial)
-        master.bind("p", self.previous_tutorial)
-        master.bind("P", self.previous_tutorial)
-        self.current_problem_index = -1
-        self.current_problem = None
-        #master.update_idletasks()
-        master.bind("<Configure>", self.resize)
-        self.allow_resize = True
-        self.current_tut_name = None
-        self.current_prob_name = None
-        #print "done init"
-
-    def update_change_tutorial_menu(self):
-        self.radio_menu.delete(0, END)
-
-        for name in self.cfg.tutorials.names:
-            self.radio_menu.add_radiobutton(label=name,
-                                            variable=self.tut_choice,
-                                            command=self.choose_tutorial)
-
-
-    def set_default_tutorial(self):
-        if self.cfg.tutorials.default != self.current_tutorial:
-            self.cfg.tutorials.default = self.current_tutorial
-
-    def remove_current_tutorial(self):
-        if self.cfg.tutorials.default == self.current_tutorial:
-            tkinter.messagebox.showerror('Remove Current Tutorial Error',
-                                         'You cannot remove the default tutorial.')
-            return
-        if len(self.cfg.tutorials.names) == 1:
-            tkinter.messagebox.showerror('Remove Current Tutorial Error',
-                                         'You cannot remove the last tutorial.')
-            return
-
-        remove_answer = tkinter.messagebox.askquestion("Remove Tutorial",
-                                                       "Do you really want to remove %s?" % self.current_tutorial)
-
-        if str(remove_answer) == 'yes':
-            del self.cfg[self.current_tutorial]
-            self.cfg.tutorials.names.remove(self.current_tutorial)
-
-            self.logout()
-            self.current_tutorial = self.cfg.tutorials.default
-            self.setup_tutorial()
-
-            self.update_change_tutorial_menu()
-
-    def setup_tutorial(self):
-        self.master.title("MyPyTutor: " + self.current_tutorial)
-        self.tut.set_directory(self.tut_dir)
-        if not self.process_tutorial_file():
-            tkinter.messagebox.showerror('Configuration Error',
-                                         'Your Tutorial folder is incorrect. Please use the Preferences menu to choose the correct folder.')
-        self.tut_interface.set_url(self.URL)
-        self.make_section_menu_entries()
-        if os.path.exists(self.ans_dir):
-            os.chdir(self.ans_dir)
-        else:
-            tkinter.messagebox.showerror('Configuration Error',
-                                         'Your Answer folder does not exist. Please use the Preferences menu to choose the folder.')
-
-    def resize(self, e):
-        if self.allow_resize:
-            self.cfg.window_sizes.problem = self.gettextlen(self.tut)
-            self.cfg.window_sizes.output = self.gettextlen(self.test_output.output)
-
-    def gettextlen(self, text_obj):
-        #print text_obj.winfo_height(), text_obj.text.dlineinfo("@0,0")
-        return text_obj.winfo_height()//text_obj.text.dlineinfo("@0,0")[3]
-
-    def update_text_lengths(self, delta):
-        #print "update_lengths"
-        self.cfg.window_sizes.problem -= delta
-        self.cfg.window_sizes.output += delta
-        self.tut.update_text_length(self.cfg.window_sizes.problem)
-        self.test_output.update_text_length(self.cfg.window_sizes.output)
-
-    def login(self):
-        if self.tut_interface.logged_on():
-            tkinter.messagebox.showerror('Login Error', 'You are already logged in.')
-            return
-        LoginDialog(self.master, self.do_login)
-        if self.tut_interface.logged_on():
-            self.toolbar.set_login(self.tut_interface.user)
-            #print self.timestamp, self.tut_interface.timestamp
-            if self.timestamp < self.tut_interface.timestamp:
-                zf = self.tut_interface.get_tut_zipfile()
-                #print zf
-                if zf is None:
-                    tkinter.messagebox.showerror('MyPyTutor', 'Error occurred when updating tutorials - try again later.')
-                    return
-                try:
-                    unzipit(zf, self.tut_dir)
-                    os.remove(zf)
-                    tkinter.messagebox.showinfo('MyPyTutor', 'Tutorial Problems Updated')
-                    self.tutorials = self.get_tutorial_info()
-                    self.make_section_menu_entries()
-                except:
-                    #print 'exception'
-                    tkinter.messagebox.showerror('MyPyTutor', 'Error occurred when updating tutorials - try again later.')
-        version = self.tut_interface.get_version()
-        if version is not None:
-            verlst = version.split('.')
-            currverlst = version_number.split('.')
-            #print currverlst
-            if verlst > currverlst:
-                if sys.version_info > (2, 6):
-                    zf = self.tut_interface.get_mpt27()
-                else:
-                    tkinter.messagebox.showerror('MyPyTutor',
-                                                 'You are not using Python 2.7 - you need to upgrade')
-                    return
-                #print zf
-                if zf is None:
-                    tkinter.messagebox.showerror('MyPyTutor',
-                                                 'Error occurred when updating MyPyTutor - try again later.')
-                    return
-                try:
-                    unzipit(zf, self.mpt_home)
-                    z = zipfile.ZipFile(zf)
-                    os.remove(zf)
-                    tkinter.messagebox.showinfo('MyPyTutor',
-                                                'MyPyTutor Updated - Please Restart')
-                except Exception as e:
-                    #print str(e)
-                    tkinter.messagebox.showerror('MyPyTutor', 'Error occurred when updating MyPyTutor - try again later.')
-        self.set_menu_colours()
-
-    def reset_menus(self):
-        self.sectionmenu.delete(0, END)
-        self.sectionmenu.add_command(label="Next Tutorial",
-                                     accelerator="N",
-                                     command=self.next_tutorial)
-        self.sectionmenu.add_command(label="Previous Tutorial",
-                                     accelerator="P",
-                                     command=self.previous_tutorial)
-
-    def do_login(self, user, passwd):
-        return self.tut_interface.login(user, passwd)
-
-    def logout(self):
-        self.toolbar.unset_login()
-        self.tut_interface.logout()
-
-    def change_password(self):
-        if self.tut_interface.logged_on():
-            cpdlg = ChangePasswordDialog(self.master)
-            self.master.wait_window(cpdlg)
-            if cpdlg.success:
-                print("Password changed")
-        else:
-            tkinter.messagebox.showerror('Change Password Error',
-                                         'You need to log in first.')
-
-    def do_change_password(self, passwd0, passwd1):
-        return self.tut_interface.change_password(passwd0, passwd1)
-
-    def upload_answer(self):
-        if self.current_problem:
-            if not self.tut_interface.logged_on():
-                self.login()
-            if not self.tut_interface.logged_on():
-                tkinter.messagebox.showerror('Upload Error', 'Not logged in.')
-                return
-            result = self.tut_interface.upload_answer(self.editor.get_text())
-            if result is None:
-                return
-            if result:
-                print("Code uploaded correctly")
-                print("WARNING: this does not submit an answer - use Submit Answer")
-            else:
-                tkinter.messagebox.showerror('Upload Error', 'Upload Error')
-
-    def download_answer(self):
-        if self.current_problem:
-            reply = self.editor.possiblysave("Save on Download Code")
-            if str(reply) == "cancel":
-                return
-            if not self.tut_interface.logged_on():
-                self.login()
-            if not self.tut_interface.logged_on():
-                tkinter.messagebox.showerror('Download Error', 'Not logged in.')
-                return
-            result = self.tut_interface.download_answer()
-            if result is None:
-                return
-            if result.startswith('Error'):
-                tkinter.messagebox.showerror('Downloadload Error', result)
-            else:
-                self.editor.preload(result)
-
-    def submit_answer(self, e=None):
-        if self.current_problem:
-            if not self.tut_interface.logged_on():
-                self.login()
-            if not self.tut_interface.logged_on():
-                tkinter.messagebox.showerror('Submit Error', 'Not logged in.')
-                return
-            result = self.tut_interface.submit_answer(self.editor.get_text())
-            if result == 'OK':
-                self.current_problem.set_status('OK')
-                self.set_menu_colours()
-                print("Answer submitted on time")
-                return
-            if result == 'LATE':
-                self.current_problem.set_status('LATE')
-                self.set_menu_colours()
-                tkinter.messagebox.showinfo('Late Submission',
-                                            'Problem submitted late.')
-            if result is None:
-                return
-            if result.startswith('Error'):
-                tkinter.messagebox.showerror('Submit Error', result)
-
-    def set_menu_colours(self):
-        result = self.tut_interface.show_submit()
-        if result is None:
-            return
-        result_dict = self.parse_results(result)
-        for tut_index, tut_name in enumerate(self.tutorials.tut_list):
-            has_not_done = False
-            has_late = False
-            tut_info = self.tutorials.tut_dict[tut_name]
-            menuobj = self.tutorials.get_menu(tut_name)
-            for index, problem_name in enumerate(tut_info.problems_list):
-                sub = result_dict.get(problem_name, '-')
-                tut_info.set_problem_status(problem_name, sub)
-                if sub == '-':
-                    menuobj.entryconfig(index, foreground='black')
-                    has_not_done = True
-                elif sub == 'OK':
-                    menuobj.entryconfig(index, foreground='blue')
-                else:
-                    menuobj.entryconfig(index, foreground='red')
-                    has_late = True
-            if has_not_done:
-                self.sectionmenu.entryconfig(tut_index+2, foreground='black')
-            elif has_late:
-                self.sectionmenu.entryconfig(tut_index+2, foreground='red')
-            else:
-                self.sectionmenu.entryconfig(tut_index+2, foreground='blue')
-
-    def show_submit(self):
-        text = str(self.tutorials)
-        TextDialog(self.master, "Results", text)
-
-    # Each student results file is of the form
-    # ##$$Name of tutorial$$##
-    # string flag: 'OK' - submitted on time 'LATE' - submitted late
-    # number of checks
-    # student code
-
-    def parse_results(self, result):
-        lines = result.split('\n')
-        result_dict = {}
-        length = len(lines)
-        i = 0
-        while i < length:
-            line = lines[i]
-            if line.startswith('##$$'):
-                header = line[4:-4]
-                result_dict[header] = lines[i+1]
-                i += 3
-            i += 1
-        return result_dict
-
-    # Create the tutorial menu based on the information in the tutorials file
-
-    def make_section_menu_entries(self):
-        self.reset_menus()
-        for index, tutname in enumerate(self.tutorials.tut_list):
-            tutinfo = self.tutorials.get_tutorial(tutname)
-            tutmenu = Menu(self.sectionmenu, tearoff=0)
-            self.sectionmenu.add_cascade(label=tutname, menu=tutmenu)
-            self.tutorials.set_menu(tutname, tutmenu)
-            for i, prob_name in enumerate(tutinfo.problems_list):
-                prob = tutinfo.problems_info[prob_name]
-                tutmenu.add_command(label=prob.name,
-                                    command=self.set_menu_command(tutname,
-                                                                  prob_name))
-
-        self.sectionmenu.add_separator()
-        self.sectionmenu.add_command(label="Close", command=self.close_event)
-        self.sectionmenu.add_command(label="Exit", command=self.close_event)
-
-    def choose_tutorial(self):
-        tut = self.tut_choice.get()
-        if self.current_tutorial != tut:
-            self.logout()
-            self.current_tutorial = tut
-
-            options = getattr(self.cfg, self.current_tutorial)
-            self.tut_dir = options.tut_dir
-            self.ans_dir = options.ans_dir
-            self.setup_tutorial()
-
-    def give_problem_feedback(self):
-        if self.current_problem:
-            FeedbackDialog(
-                self.master,
-                "Problem Feedback: {}".format(self.current_problem.name),
-                self.current_problem,
-                self.editor.get_text()
-            )
-
-    def give_general_feedback(self):
-        FeedbackDialog(self.master, 'General Feedback', '')
-
-    def configure_tut(self):
-        result = FontChooser(
-            self.master, self, (self.cfg.font.name, self.cfg.font.size),
-        ).result
-        if result:
-            self.cfg.font.name = result[0]
-            self.cfg.font.size = result[1]
-
-            self.tut.update_fonts(result[0], int(result[1]))
-            self.test_output.update_font(int(result[1]))
-
-    def configure_tut_dir(self):
-        if os.path.exists(self.tut_dir):
-            initial_dir = self.tut_dir
-        else:
-            initial_dir = os.path.expanduser('~')
-
-        dir = tkinter.filedialog.askdirectory(
-            title='Choose Tutorial Folder: {}'.format(self.current_tutorial),
-            initialdir=initial_dir,
-        )
-        if dir:
-            options = getattr(self.cfg, self.current_tutorial)
-            options.tut_dir = dir
-
-            self.tut_dir = dir
-
-            if self.sections:
-                self.sectionmenu.delete(2, self.sections+6)
-            if not self.process_tutorial_file():
-                tkinter.messagebox.showerror('Configuration Error',
-                                             'Your Tutorial folder is incorrect. Please use the Preferences menu to choose the correct folder.')
-            self.tut_interface.set_url(self.URL)
-            self.make_section_menu_entries()
-
-    def configure_answers_dir(self):
-        if os.path.exists(self.ans_dir):
-            initial_dir = self.ans_dir
-        else:
-            initial_dir = os.path.expanduser('~')
-
-        dir = tkinter.filedialog.askdirectory(
-            title='Choose Answers Folder: {}'.format(self.current_tutorial),
-            initialdir=initial_dir,
-        )
-        if dir:
-            options = getattr(self.cfg, self.current_tutorial)
-            options.ans_dir = dir
-
-            self.ans_dir = dir
-
-            os.chdir(self.ans_dir)
-
-    def add_new_tutorial(self):
-        error = add_tutorial(self.cfg, window=self.master, as_default=False)
-
-        if error:
-            tkinter.messagebox.showerror('Add New Tutorial Error', error)
-            return
-
-        self.update_change_tutorial_menu()
-
-    def font_apply(self, font_name, font_size):
-        self.tut.update_fonts(font_name, int(font_size))
-        if self.test_output:
-            self.test_output.update_font(int(font_size))
-        if self.editor is not None:
-            self.editor.update_font(font_size)
-
-    def next_tutorial(self, e=None):
-        self.menu_choice(*self.tutorials.next_tutorial(self.current_tut_name,
-                                                       self.current_prob_name))
-
-    def previous_tutorial(self, e=None):
-        if (self.current_tut_name is not None and
-                self.current_prob_name is not None):
-            self.menu_choice(*self.tutorials.previous_tutorial(self.current_tut_name,
-                                                               self.current_prob_name))
-
-    def set_menu_command(self, tut_name, prob_name):
-        return lambda: self.menu_choice(tut_name, prob_name)
-
-    def menu_choice(self, tut_name, problem_name):
-        self.current_tut_name = tut_name
-        self.current_prob_name = problem_name
-        online = self.URL is not None
-        if self.editor is None:
-            ## Create a code editor window if none
-            self.editor = tut_editor.TutorEditor(self, root=self.master,
-                                                 online=online)
-            #print "after create edit window"
-            self.tut_interface.set_editor(self.editor)
-            self.font_apply(self.cfg.font.name, self.cfg.font.size)
-        if self.editor.maybesave() == "cancel":
-            return
-        answer_file = os.path.join(self.ans_dir,
-                                   '_'.join(problem_name.split()) + '.py')
-        self.master.title(problem_name)
-        self.current_problem = self.tutorials.get_problem(tut_name,
-                                                          problem_name)
-        self.tut_interface.load_data(os.path.join(self.tut_dir,
-                                                  self.current_problem.tut),
-                                     problem_name)
-        if self.tut_interface.get_text():
-            try:
-                self.tut.add_text(self.tut_interface.get_text())
-                self.tut_interface.reset_editor(answer_file)
-                self.editor.undo.reset_undo()
-                self.toolbar.set_hints(self.tut_interface.get_hints())
-            except Exception as e:
-                print('Exception: ' + str(e), file=sys.stderr)
-
-        self.short_description.config(
-            text=self.tut_interface.get_short_description()
-        )
-
-        # immediately run the tests; this will update the display
-        self.run_tests()
-
-    def get_tutorial_info(self):
-        return tut_tutorial.TutorialInfo(self.tut_dir)
-
-    def process_tutorial_file(self):
-        info = self.get_tutorial_info()
-        if info is None:
-            self.tutorials = []
-            return False
-        self.tutorials = info
-        self.URL = None
-        try:
-            config_file = os.path.join(self.tut_dir, 'config.txt')
-            fid = open(config_file, 'U')
-            text = fid.read()
-            fid.close()
-            lines = text.split('\n')
-            self.timestamp = float(lines[0])
-            self.URL = lines[1]
-        except:
-            return False
-        return True
-
-    ## Run the tests on the user code
-
-    def run_tests(self):
-        tester, analyser = self.tut_interface.run_tests(self.editor.get_text())
-
-        self.test_output.set_test_results(tester.results)
-        self.analysis_output.set_analyser(analyser)
-
-    def get_preloaded(self):
-        return self.tut_interface.get_preloaded()
-
-    def showhint(self):
-        hint = self.tut_interface.get_next_hint()
-        if hint:
-            self.tut.show_hint(hint)
-
-    def close_event(self, _e=None):
-        if self.editor is not None:
-            result = self.editor.close()
-            if str(result) == 'yes':
-                self.master.destroy()
-        else:
-            self.master.destroy()
-
-        self.logout()
-
-        save_config(self.cfg)
-
-    def quit_editor(self):
-        self.editor = None
-
-    def about(self):
-        TutAboutDialog(self.master, "About MyPyTutor")
-
-    def help(self):
-        HelpDialog(self.master, "Help")
